Refactor.
- Update Target Framework of the unit test project with .NET Core 3.1.
