+++ .circleci/config.yml
@@ -2,7 +2,7 @@ version: 2
 jobs:
   build:
     docker:
-      - image: mcr.microsoft.com/dotnet/core/sdk:3.0.100
+      - image: mcr.microsoft.com/dotnet/core/sdk:3.1.201
     steps:
       - run:
           name: Register Microsoft key and feed
